Fixed coding style from conductance.py

+++ src/aihwkit/inference/converter/conductance.py
@@ -2,21 +2,18 @@
 
 # (C) Copyright 2020, 2021, 2022, 2023, 2024 IBM. All Rights Reserved.
 #
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
+# Licensed under the MIT license. See LICENSE file in the project root for details.
 
 """Conductance converters for the phenomenological noise models for inference."""
 
 from typing import Dict, List, Optional, Tuple
 
-from torch import Tensor, zeros_like, from_numpy, linspace, allclose, stack, abs as torch_abs
+from torch import abs as torch_abs
+from torch import Tensor, zeros_like, from_numpy, linspace, allclose
 from torch.autograd import no_grad
+
 from numpy import interp
+
 from aihwkit.inference.converter.base import BaseConductanceConverter
 
 _ZERO_CLIP = 1e-7
@@ -68,17 +65,15 @@ class SinglePairConductanceConverter(BaseConductanceConverter):
         return conductances, params
 
     @no_grad()
-    def convert_back_to_weights(
-        self, conductances: List[Tensor], params: Dict
-    ) -> Tensor:
+    def convert_back_to_weights(self, conductances: List[Tensor], params: Dict) -> Tensor:
         if len(conductances) != 2:
             raise ValueError("conductances must contain exactly two elements")
         if "scale_ratio" not in params:
             raise ValueError("params do not contain scale_ratio")
 
-        weights = (
-            (conductances[0] - self.g_min) - (conductances[1] - self.g_min)
-        ) / params["scale_ratio"]
+        weights = ((conductances[0] - self.g_min) - (conductances[1] - self.g_min)) / params[
+            "scale_ratio"
+        ]
 
         return weights
 
@@ -108,17 +103,16 @@ class NPairConductanceConverter(BaseConductanceConverter):
             the logical zero of the weights will be mapped to.
     """
 
-    def __init__(
-        self,
-        f_lst: List[float],
-        g_max: Optional[float] = None,
-        g_min: Optional[float] = None,
-    ):
+    def __init__(self,
+                 f_lst: List[float],
+                 g_max: Optional[float] = None,
+                 g_min: Optional[float] = None,
+                 ):
 
         if not isinstance(f_lst, list):
             raise ValueError("f_lst parameter must be a list of F factors")
 
-        if max(f_lst) < 0.0:
+        if max(f_lst) < 0.:
             raise ValueError("f_lst parameter contains negative value")
 
         self.g_max = 25.0 if g_max is None else g_max
@@ -145,44 +139,35 @@ class NPairConductanceConverter(BaseConductanceConverter):
         scale_ratio = max_weight_us / max_weight_unitless
         weights_us = scale_ratio * weights
 
-        lower_bound_us = 0.0  # lower bound in uS
+        lower_bound_us = 0.  # lower bound in uS
         conductances = []
         for f_factor in self.f_lst:
-            conductances.append(
-                (
-                    (weights_us.clamp(min=0.0) - lower_bound_us) / f_factor + self.g_min
-                ).clamp(min=self.g_min, max=self.g_max)
-            )  # g_plus
-            conductances.append(
-                (
-                    ((-weights_us).clamp(min=0.0) - lower_bound_us) / f_factor
-                    + self.g_min
-                ).clamp(min=self.g_min, max=self.g_max)
-            )  # g_minus
+            conductances.append(((weights_us.clamp(min=0.0) - lower_bound_us) / f_factor
+                                 + self.g_min).clamp(min=self.g_min, max=self.g_max))      # g_plus
+            conductances.append((((-weights_us).clamp(min=0.0) - lower_bound_us) / f_factor
+                                 + self.g_min).clamp(min=self.g_min, max=self.g_max))      # g_minus
             lower_bound_us += f_factor * (self.g_max - self.g_min)
 
-        params = {"scale_ratio": scale_ratio, "f_lst": self.f_lst}
+        params = {'scale_ratio': scale_ratio,
+                  'f_lst': self.f_lst}
 
         return conductances, params
 
     @no_grad()
-    def convert_back_to_weights(
-        self, conductances: List[Tensor], params: Dict
-    ) -> Tensor:
-        if "f_lst" not in params:
+    def convert_back_to_weights(self, conductances: List[Tensor], params: Dict) -> Tensor:
+        if 'f_lst' not in params:
             raise ValueError("params does not contain f_lst")
         if len(conductances) % 2 != 0:
             raise ValueError("unit cell must have an even number of conductances")
-        if "scale_ratio" not in params:
+        if 'scale_ratio' not in params:
             raise ValueError("params does not contain scale_ratio")
 
         weights = zeros_like(conductances[0])
-        for f_factor, (g_plus, g_minus) in zip(
-            self.f_lst, zip(conductances[::2], conductances[1::2])
-        ):
+        for f_factor, (g_plus, g_minus) in zip(self.f_lst,
+                                               zip(conductances[::2], conductances[1::2])):
             weights += f_factor * (g_plus - g_minus)
 
-        return weights / params["scale_ratio"]
+        return weights / params['scale_ratio']
 
 
 class DualPairConductanceConverter(NPairConductanceConverter):
@@ -210,17 +195,18 @@ class DualPairConductanceConverter(NPairConductanceConverter):
             the logical zero of the weights will be mapped to.
     """
 
-    def __init__(
-        self,
-        f_lst: List[float],
-        g_max: Optional[float] = None,
-        g_min: Optional[float] = None,
-    ):
+    def __init__(self,
+                 f_lst: List[float],
+                 g_max: Optional[float] = None,
+                 g_min: Optional[float] = None,
+                 ):
 
         if len(f_lst) != 2:
             raise ValueError("f_lst parameter does not contain two values")
 
-        super().__init__(f_lst=f_lst, g_max=g_max, g_min=g_min)
+        super().__init__(f_lst=f_lst,
+                         g_max=g_max,
+                         g_min=g_min)
 
 
 class CustomPairConductanceConverter(BaseConductanceConverter):
@@ -249,14 +235,13 @@ class CustomPairConductanceConverter(BaseConductanceConverter):
             the logical zero of the weights will be mapped to.
     """
 
-    def __init__(
-        self,
-        f_lst: List[float],
-        g_lst: List[List[float]],
-        g_max: Optional[float] = None,
-        g_min: Optional[float] = None,
-        invertibility_test: Optional[bool] = True,
-    ):
+    def __init__(self,
+                 f_lst: List[float],
+                 g_lst: List[List[float]],
+                 g_max: Optional[float] = None,
+                 g_min: Optional[float] = None,
+                 invertibility_test: Optional[bool] = True,
+                 ):
         self.g_max = 25.0 if g_max is None else g_max
         self.g_min = 0.0 if g_min is None else g_min
 
@@ -273,9 +258,7 @@ class CustomPairConductanceConverter(BaseConductanceConverter):
             raise ValueError("f_lst parameter must be a list")
 
         if 2 * len(f_lst) != len(g_lst):
-            raise ValueError(
-                "must have one value in f_lst for every pair of values in g_lst"
-            )
+            raise ValueError("must have one value in f_lst for every pair of values in g_lst")
 
         self.f_lst = f_lst
         self.g_lst = g_lst
@@ -306,9 +289,7 @@ class CustomPairConductanceConverter(BaseConductanceConverter):
         conductances, params = self.convert_to_conductances(test_weights)
         return_weights = self.convert_back_to_weights(conductances, params)
         if not allclose(test_weights, return_weights, atol=0.0001):
-            raise ArithmeticError(
-                "CustomPairConductanceConverter is not an invertible function"
-            )
+            raise ArithmeticError("CustomPairConductanceConverter is not an invertible function")
 
     def __str__(self) -> str:
         return "{}(g_max={:1.2f}, g_min={:1.2f})".format(
@@ -319,62 +300,51 @@ class CustomPairConductanceConverter(BaseConductanceConverter):
     def convert_to_conductances(self, weights: Tensor) -> Tuple[List[Tensor], Dict]:
 
         weights_us = zeros_like(Tensor(self.g_lst[0])).type_as(weights)
-        for f_factor, (gp_lst, gm_lst) in zip(
-            self.f_lst, zip(self.g_lst[::2], self.g_lst[1::2])
-        ):
+        for f_factor, (gp_lst, gm_lst) in zip(self.f_lst,
+                                              zip(self.g_lst[::2], self.g_lst[1::2])):
             weights_us += f_factor * (Tensor(gp_lst) - Tensor(gm_lst)).type_as(weights)
 
         max_weight = torch_abs(weights).max()
         max_weight_us = torch_abs(weights_us).max()
         scale_ratio = max_weight_us / max_weight.clamp(min=_ZERO_CLIP)
 
-        w_lst = (
-            linspace(-max_weight_us, max_weight_us, len(self.g_lst[0])) / scale_ratio
-        ).tolist()
+        w_lst = (linspace(-max_weight_us,
+                          max_weight_us,
+                          len(self.g_lst[0])) / scale_ratio).tolist()
 
         conductances = []
-        for f_factor, (gp_lst, gm_lst) in zip(
-            self.f_lst, zip(self.g_lst[::2], self.g_lst[1::2])
-        ):
-            conductances.append(
-                from_numpy(interp(weights.cpu().numpy(), w_lst, gp_lst)).type_as(
-                    weights
-                )
-            )
-            conductances.append(
-                from_numpy(interp(weights.cpu().numpy(), w_lst, gm_lst)).type_as(
-                    weights
-                )
-            )
+        for f_factor, (gp_lst, gm_lst) in zip(self.f_lst,
+                                              zip(self.g_lst[::2], self.g_lst[1::2])):
+            conductances.append(from_numpy(interp(weights.cpu().numpy(),
+                                                  w_lst,
+                                                  gp_lst)).type_as(weights))
+            conductances.append(from_numpy(interp(weights.cpu().numpy(),
+                                                  w_lst,
+                                                  gm_lst)).type_as(weights))
 
-        params = {"scale_ratio": scale_ratio, "f_lst": self.f_lst}
+        params = {'scale_ratio': scale_ratio,
+                  'f_lst': self.f_lst}
 
         return conductances, params
 
     @no_grad()
-    def convert_back_to_weights(
-        self, conductances: List[Tensor], params: Dict
-    ) -> Tensor:
+    def convert_back_to_weights(self, conductances: List[Tensor], params: Dict) -> Tensor:
 
-        if "f_lst" not in params:
+        if 'f_lst' not in params:
             raise ValueError("params does not contain f_lst")
 
-        if not isinstance(params["f_lst"], list):
+        if not isinstance(params['f_lst'], list):
             raise TypeError("f_lst parameter must be a list of f factors")
 
-        if 2 * len(params["f_lst"]) != len(conductances):
-            raise ValueError(
-                "must have one value in f_lst for every pair of conductances"
-            )
+        if 2 * len(params['f_lst']) != len(conductances):
+            raise ValueError("must have one value in f_lst for every pair of conductances")
 
         weights_us = zeros_like(conductances[0])
-        for f_factor, (g_plus, g_minus) in zip(
-            params["f_lst"], zip(conductances[::2], conductances[1::2])
-        ):
+        for f_factor, (g_plus, g_minus) in zip(params['f_lst'],
+                                               zip(conductances[::2], conductances[1::2])):
             weights_us += f_factor * (g_plus - g_minus)
 
-        return weights_us / params["scale_ratio"]  # back to unitless
-
+        return weights_us / params['scale_ratio']   # back to unitless
 
 class SingleDeviceConductanceConverter(BaseConductanceConverter):
     r"""Single devices to represent weights
